Lesson changes, deleted unnecessary files, added questionares for blocks.
