+++ src/pages/LessonPageEugene.tsx
@@ -87,7 +87,7 @@ const LessonPage = () => {
               Lessons
             </Text>
             <Box display="flex" flexDir="column" gap="30px">
-              {lessons.map((element) => {
+              {lessons.map((element, index) => {
                 return (
                   <Box display="flex" alignItems="center">
                     <Box mr="10px">
